backend/requirements.txt
# Core backend framework
flask==3.0.3
flask-cors==5.0.0

# Database ORM
sqlalchemy==2.0.36
psycopg2-binary==2.9.9   # PostgreSQL driver

# Data handling
pandas==2.2.3
numpy==1.26.4

# Optimization & AI
scikit-learn==1.5.2
scipy==1.14.1
ortools==9.10.4067        # for optimization engine (LP/MIP)

# Utils
python-dotenv==1.0.1

flask-sqlalchemy==3.1.1
